+++ python/numpy/random.py
@@ -1,0 +1,13 @@
+#!/bin/env python
+
+import numpy as np
+
+true_b = 1
+true_w = 2
+N = 100
+
+# Data Generation
+np.random.seed(42)
+x = np.random.rand(N, 1)
+epsilon = (.1 * np.random.randn(N, 1))
+y = true_b + true_w * x + epsilon