+++ glibc/pthread/samples/condattr_setclock.c
@@ -1,0 +1,1 @@
+condattr_getclock.c