+++ glibc/pthread/samples/getconcurrency.c
@@ -1,0 +1,1 @@
+setconcurrency.c