+++ glibc/pthread/samples/mutexattr_setkind_np.c
@@ -1,0 +1,1 @@
+mutexattr_settype.c