+++ glibc/pthread/samples/pthread_setname_np.c
@@ -1,1 +1,0 @@
-pthread_getname_np.c