+++ glibc/pthread/samples/setschedparam.c
@@ -1,0 +1,1 @@
+getschedparam.c