+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to
 #### Added
 - python: add numpy;
 - pthread: move `pthread_helpers` to top pthread directory;
-- ai: Introduce tensorflow;
+- ai: Introduce tensorflow, mnist;
 #### Changed
 #### Deprecated
 #### Removed