+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
-- python/numpy: add array directory;
+- python/numpy: add array test;
 - keras: Train MNIST with keras;
 - python: add numpy;
 - pthread: move `pthread_helpers` to top pthread directory;